3.2.8

+++ scripts/deleteAlias.php
@@ -12,40 +12,50 @@ require 'wampserver.lib.php';
 
 $alias = $_SERVER['argv'][1];
 $aliasToDelete = $alias.".conf";
+$aliasToDelete = str_replace('-whitespace-',' ',$aliasToDelete);
 $message = '';
+$deleteAlias = $deleteDir = false;
 if(file_exists($aliasDir.$aliasToDelete)) {
-	$message .= "Alias : ".$aliasToDelete."\n\nDo you really want to delete this alias?\nType Y key for yes: ";
+	$message .= "Alias : ".$aliasDir.$aliasToDelete."\n\nDo you really want to delete this alias?\nType Y key then Enter for yes: ";
 	Command_Windows($message,80,-1,0,'delete an Alias');
 	$confirm = trim(fgets(STDIN));
 	$confirm = strtoupper(trim($confirm ,'\''));
 	if($confirm == 'Y') {
+		$deleteAlias = true;
 		$alias_contents = @file_get_contents($aliasDir.$aliasToDelete);
-	  preg_match('~^Alias\s+/'.$alias.'\s+"(.*)"\r?$~m',$alias_contents,$matches);
-	  if(is_dir($matches[1])) {
-	  	$deleteDir = false;
-	  	$message .= "\n\nThe directory ".$matches[1]." is associated with alias ".$alias."\n\nDo you really want to delete also this directory?\n\nType Y key for yes: ";
-			Command_Windows($message,80,-1,0,'Delete an Alias');
-			$confirm = trim(fgets(STDIN));
-			$confirm = strtoupper(trim($confirm ,'\''));
-			if($confirm == 'Y') {
-				$deleteDir = true;
-				$dirToDelete = $matches[1];
-			}
-	  }
-		if(unlink($aliasDir.str_replace('-whitespace-',' ',$aliasToDelete))) {
-			$message .= "\n\n\nAlias deleted\n";
-			if($deleteDir) {
-				if(file_exists($dirToDelete) && is_dir($dirToDelete)) {
-					if(rrmdir($dirToDelete) === false)
-						$message .= "\n\nFolder ".$dirToDelete." **not** deleted\n";
-					else
-						$message .= "\n\nFolder ".$dirToDelete." deleted\n";
+	  if(preg_match('~^Alias\s+/'.$alias.'\s+"(.*)"\r?$~m',$alias_contents,$matches) === 1){
+	  	$dirToDelete = replace_apache_var($matches[1]);
+	  	if(is_dir($dirToDelete)) {
+	  		$deleteDir = false;
+	  		$message .= "\n\nThe directory ".$dirToDelete."\nis associated with alias ".$aliasDir.$aliasToDelete."\nDo you really want to delete also this directory?\n\nType Y key then Enter for yes: ";
+	  		$message1 = "\n\nThe directory ".$dirToDelete."\nis associated with alias ".$aliasDir.$aliasToDelete."\n\nDo you really want to delete also this directory?\n\nType Y key then Enter for yes: ";
+				//Command_Windows($message,80,-1,0,'Delete an Alias');
+				echo $message1;
+				$confirm = trim(fgets(STDIN));
+				$confirm = strtoupper(trim($confirm ,'\''));
+				if($confirm == 'Y') {
+					$deleteDir = true;
+				}
+	  	}
+		}
+	  $message = '';
+	  if($deleteAlias) {
+			if(unlink($aliasDir.str_replace('-whitespace-',' ',$aliasToDelete))) {
+				$message .= color('blue',"\nAlias ".$aliasDir.$aliasToDelete." deleted\n");
+				if($deleteDir) {
+					if(file_exists($dirToDelete) && is_dir($dirToDelete)) {
+						if(rrmdir($dirToDelete) === false)
+							$message .= color('red',"\n\nFolder ".$dirToDelete." **not** deleted\n");
+						else
+							$message .= color('blue',"\nFolder ".$dirToDelete." deleted\n");
+					}
 				}
 			}
+			else
+				$message .= color('red',"\n\n**Unable to delete** alias with unlink()\n");
 		}
-		else
-			$message .= "\n\n**Unable to delete** alias with unlink()\n";
-		$message .=  "\n\nPress Enter to exit ";
+
+		$message .=  "\nPress Enter to exit ";
 		Command_Windows($message,80,-1,0,'Delete an Alias');
 		trim(fgets(STDIN));
 	  exit();
